requirements.txt
autopep8==1.4.3
chardet==3.0.4
ChatterBot==0.8.6
chatterbot-corpus==1.1.4
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
future==0.17.1
get==2019.3.22
idna==2.8
mathparse==0.1.2
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
opencc-python-reimplemented==0.1.4
post==2019.3.22
public==2019.3.22
pycodestyle==2.5.0
pymongo==3.7.2
python-dateutil==2.6.1
python-twitter==3.5
pytz==2018.9
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
SQLAlchemy==1.2.18
urljoin==1.0.0
urllib3==1.24.1